+++ app/src/main/java/de/uniks/codliners/stock_simulator/repository/AccountRepository.kt
@@ -36,7 +36,7 @@ class AccountRepository(private val database: StockAppDatabase) {
     fun depotQuoteWithSymbol(symbol: String): LiveData<DepotQuote> =
         database.accountDao.getDepotQuoteWithSymbol(symbol)
 
-    suspend fun buy(quote: Quote, amount: Int) {
+    suspend fun buy(quote: Quote, amount: Double) {
         val lastBalance = latestBalance.value
         lastBalance?.let {
             withContext(Dispatchers.IO) {
@@ -44,7 +44,7 @@ class AccountRepository(private val database: StockAppDatabase) {
                 val newBalance = Balance(lastBalance.value + cashflow)
 
                 val depotQuote = database.accountDao.getDepotQuoteBySymbol(quote.symbol)
-                    ?: DepotQuote(quote.symbol, 0)
+                    ?: DepotQuote(quote.symbol, 0.0)
                 val newDepotQuote = depotQuote.copy(amount = depotQuote.amount + amount)
 
                 val transaction = TransactionDatabase(
@@ -67,7 +67,7 @@ class AccountRepository(private val database: StockAppDatabase) {
         }
     }
 
-    suspend fun sell(quote: Quote, amount: Int) {
+    suspend fun sell(quote: Quote, amount: Double) {
         val lastBalance = latestBalance.value
         lastBalance?.let {
             withContext(Dispatchers.IO) {
